Moved out CellState, PlayerID and PlayerMove to separate files
