+++ game/player_move.go
@@ -1,0 +1,8 @@
+package game
+
+import "github.com/kitsunemikan/six-purrpurrs/geom"
+
+type PlayerMove struct {
+	Cell geom.Offset
+	ID   PlayerID
+}
